[MINOR] Move affinityFieldName to GridCacheUtils
