+++ modules/core/src/main/java/org/apache/ignite/internal/binary/BinaryContext.java
@@ -19,7 +19,6 @@ package org.apache.ignite.internal.binary;
 
 import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.math.BigDecimal;
 import java.net.URISyntaxException;
 import java.net.URL;
@@ -61,7 +60,6 @@ import org.apache.ignite.binary.BinaryReflectiveSerializer;
 import org.apache.ignite.binary.BinarySerializer;
 import org.apache.ignite.binary.BinaryType;
 import org.apache.ignite.binary.BinaryTypeConfiguration;
-import org.apache.ignite.cache.affinity.AffinityKeyMapped;
 import org.apache.ignite.internal.DuplicateTypeIdException;
 import org.apache.ignite.internal.UnregisteredBinaryTypeException;
 import org.apache.ignite.internal.UnregisteredClassException;
@@ -141,6 +139,9 @@ public class BinaryContext {
     /** */
     private final Function<String, BinaryInternalMapper> mapperProvider;
 
+    /** */
+    private final Function<Class<?>, String> affFldNameProvider;
+
     /** Logger. */
     private final IgniteLogger log;
 
@@ -163,6 +164,7 @@ public class BinaryContext {
      * @param nameMapper Binary name mapper.
      * @param affFlds Affinity fields.
      * @param compactFooter Compact footer flag.
+     * @param affFldNameProvider Affinity field name provider function.
      * @param log Logger.
      */
     public BinaryContext(
@@ -176,6 +178,7 @@ public class BinaryContext {
         @Nullable Collection<BinaryTypeConfiguration> typeCfgs,
         Map<String, String> affFlds,
         boolean compactFooter,
+        Function<Class<?>, String> affFldNameProvider,
         IgniteLogger log
     ) {
         assert metaHnd != null;
@@ -193,6 +196,7 @@ public class BinaryContext {
             mapperProvider = clsName -> DFLT_MAPPER;
 
         this.compactFooter = compactFooter;
+        this.affFldNameProvider = affFldNameProvider;
 
         this.log = log;
 
@@ -371,7 +375,7 @@ public class BinaryContext {
                             Class<?> cls = U.classForName(clsName0, null);
 
                             if (cls != null)
-                                affField = affinityFieldName(cls);
+                                affField = affFldNameProvider.apply(cls);
                         }
 
                         descs.add(clsName0, mapper, serializer, identity, affField,
@@ -385,7 +389,7 @@ public class BinaryContext {
                         Class<?> cls = U.classForName(clsName, null);
 
                         if (cls != null)
-                            affField = affinityFieldName(cls);
+                            affField = affFldNameProvider.apply(cls);
                     }
 
                     descs.add(clsName, mapper, serializer, identity, affField,
@@ -658,7 +662,7 @@ public class BinaryContext {
             BinarySerializer serializer = serializerForClass(cls);
 
             // Firstly check annotations, then check in cache key configurations.
-            String affFieldName = affinityFieldName(cls);
+            String affFieldName = affFldNameProvider.apply(cls);
             if (affFieldName == null)
                 affFieldName = affKeyFieldNames.get(typeId);
 
@@ -785,7 +789,7 @@ public class BinaryContext {
 
         BinarySerializer serializer = serializerForClass(cls);
 
-        String affFieldName = affinityFieldName(cls);
+        String affFieldName = affFldNameProvider.apply(cls);
 
         return new BinaryClassDescriptor(this,
             cls,
@@ -1045,21 +1049,6 @@ public class BinaryContext {
         BinaryInternalMapper mapper = userTypeMapper(clsName);
 
         return mapper.typeName(clsName);
-    }
-
-    /**
-     * @param cls Class to get affinity field for.
-     * @return Affinity field name or {@code null} if field name was not found.
-     */
-    public static String affinityFieldName(Class cls) {
-        for (; cls != Object.class && cls != null; cls = cls.getSuperclass()) {
-            for (Field f : cls.getDeclaredFields()) {
-                if (f.getAnnotation(AffinityKeyMapped.class) != null)
-                    return f.getName();
-            }
-        }
-
-        return null;
     }
 
     /**
