+++ modules/core/src/test/java/org/apache/ignite/testframework/junits/IgniteMock.java
@@ -459,6 +459,7 @@ public class IgniteMock implements IgniteEx {
                 bcfg.getTypeConfigurations(),
                 CU.affinityFields(configuration()),
                 bcfg.isCompactFooter(),
+                CU::affinityFieldName,
                 NullLogger.INSTANCE
             ) {
                 @Override public int typeId(String typeName) {
